Implement Phase 12 — V5 Governance integration
- House private construction remains a direct House action without governance requirement
- Corporation public construction requires unified governance proposal authorization
- Corporation scale capability research integrated with Corporation governance
- Earth technology frontier advancement integrated with Earth governance; legacy technology mutation restricted
- Double-entry accounting for Credit and resources across Treasury and Inventory to settlement and consumption sinks
- Unified canonical proposal lifecycle preserved with frozen rules, electorate, voting, and activation provenance
- Added migration 128_v5_governance_integration.sql and updated schema manifest/contract

+++ cloudflare/src/earth-technology-frontier-postgres.ts
@@ -101,7 +101,7 @@ export async function advanceEarthTechnologyFrontier(
     input.proposalId, input.humanId ?? null, input.correlationId,
   ]);
   await tx.query(`UPDATE earth_technology_frontier
-                     SET max_generation_number = $2, updated_game_day = $3, updated_at = CURRENT_TIMESTAMP
-                   WHERE domain_id = $1 AND $3 >= $4`, [input.domainId, input.generationNumber, executionGameDay, input.effectiveFromGameDay]);
+                     SET max_generation_number = $2, updated_game_day = $3::BIGINT, updated_at = CURRENT_TIMESTAMP
+                   WHERE domain_id = $1 AND $3::BIGINT >= $4::BIGINT`, [input.domainId, input.generationNumber, executionGameDay, input.effectiveFromGameDay]);
   return { ok: true, frontierVersionId: id, domainId: input.domainId, generationNumber: input.generationNumber, effectiveFromGameDay: input.effectiveFromGameDay, correlationId: input.correlationId };
 }
